ML/requirements.txt
flask==3.1.0
flask-cors==5.0.1
tensorflow==2.16.1
numpy
pillow
gradio
